client and owner agreement structure

+++ components/Authentication/Login.js
@@ -1,10 +1,11 @@
-import React from "react";
+import React, { useState } from "react";
 import { Formik, useFormik } from "formik";
 import * as Yup from "yup";
 import {
   Button,
   FormControl,
   FormErrorMessage,
+  HStack,
   Stack,
   useTimeout,
   useToast,
@@ -17,6 +18,7 @@ import { useUserStore } from "../../stores/user.store";
 import { useRouter } from "next/router";
 import OverviewServices from "../../services/overview.service";
 import LocalStorageService from "../../services/localstorage.service";
+import { useOwnerBasket } from "../../stores/ownerBasket.store";
 
 const LoginSchema = Yup.object({
   name: Yup.string()
@@ -29,13 +31,15 @@ const LoginSchema = Yup.object({
 
 function Login({ closeModal }) {
 
+  const [user, setUser] = useState("");
   const router = useRouter();
   const currentUrl = router.asPath;
   const toast = useToast();
-  const {setUserData, userData, clearUserData } = useUserStore((state) => state);
+  const {setUserData, userData, clearUserData, setClientData } = useUserStore((state) => state);
+  const { ownerBasket, setOwnerBasket, clearOwnerBasket} = useOwnerBasket((state) => state)
 
 
-  const { mutate: dataForLogin } = useMutation(AuthenticationService.loginClientWithUsername, {
+  const { mutate: dataForClientLogin } = useMutation(AuthenticationService.loginClientWithUsername, {
     onSuccess: async (data) => {
       clearUserData();
       LocalStorageService.remove("jwtToken");
@@ -75,14 +79,73 @@ function Login({ closeModal }) {
 }
 )
 
+const { mutate: dataForOwnerLogin} = useMutation(AuthenticationService.loginOwnerWithUsername, {
+  onSuccess: async (data) => {
+      clearUserData();
+      LocalStorageService.remove("jwtToken");
+      await new Promise((resolve) => {
+        // Ensure data is valid
+              if (data && data.token) {
+                console.log(data)
+                setUserData({ jwtToken: data.token });
+                LocalStorageService.set("jwtToken", data.token);
+              } else {
+                console.error("Invalid data received");
+              }
+          
+            resolve(); 
+          });
+            toast({
+              title: 'Login Successful',
+              description: `Driving To Overview.....`,
+              status: 'success',
+              duration: 2000,
+              position: 'top',
+              isClosable: true,
+            })
+            router.push('/ownerOverview');
+  },
+  onError: (error) => {
+    console.log(error);
+    toast({
+      title: 'Login un-Successful',
+      description: "You are not logedIn now.",
+      status: 'error',
+      duration: 2000,
+      position: 'top',
+      isClosable: true,
+    })
+  }
+})
+
   const handleLogin = async (formdata) => {
+    if (user === "CLIENT") {
+    setClientData({
+      clientName: formdata.name
+    });
     const data = {
       name: formdata.name,
       password: formdata.password
     }
+    if(LocalStorageService.get('jwtToken')){
     LocalStorageService.remove("jwtToken");
-    dataForLogin(data);
+    }
+    dataForClientLogin(data);
     closeModal();
+  } else if (user === "OWNER") {
+    setOwnerBasket({
+      ownerName: formdata.name
+    });
+    const data = {
+      name: formdata.name,
+      password: formdata.password
+  }
+  if(LocalStorageService.get('jwtToken')){
+    LocalStorageService.remove("jwtToken");
+    }
+  dataForOwnerLogin(data);
+  closeModal();
+}
     // Api to check whether user is present or not; use formdata.name, formdata.password
   };
 
@@ -113,6 +176,16 @@ function Login({ closeModal }) {
             <PasswordInput {...getFieldProps("password")} />
             <FormErrorMessage>{errors.password}</FormErrorMessage>
           </FormControl>
+          <HStack>
+            <Button onClick={() => setUser("CLIENT")} 
+              bg={user == "CLIENT" ? "yellow.400" : "gray.200"}
+              color={"black"}
+              >Client</Button>
+            <Button onClick={() => setUser("OWNER")}
+              bg={user == "OWNER" ? "yellow.400" : "gray.200"}
+              color={"black"}
+              >Owner</Button>
+            </HStack>
           <Stack direction={["column", "row"]} w="full" spacing={5}>
           <Button type="submit" variant="primary" color={"#FFD76F"} bg={"grey"}>
               Login
